// src/types/TerraceType.ts
export type Terrace = {
  id: string;
  business_name: string;
  cadastro_ref: string;
  address: string;
  activity_code: number;
  group_activity_code: number;
  district_name: string;
  district_code: number;
  neighbourhood_name: string;
  neighbourhood_code: number;
  zip_code: string;
  tables: number;
  seats: number;
  latitude: number;
  longitude: number;
  average_price: number | null;
  average_rating: number;
  has_wifi: boolean | null;
  pet_friendly: boolean | null;
  can_smoke: boolean | null;
  has_disabled_access: boolean | null;
  has_kitchen: boolean | null;
  reservation_fee: number;
  is_claimed: boolean;
  instagram_account: string;
  website: string;
  profile_pic: string;
};